include container comaparison in logical backup diff
